feat: renamed database package to model
