+++ partshop-backend/src/main/java/hu/kosztaauto/partshop/repository/RaktarRepository.java
@@ -1,9 +1,7 @@
 package hu.kosztaauto.partshop.repository;
 
-import hu.kosztaauto.partshop.database.Raktar;
+import hu.kosztaauto.partshop.model.Raktar;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public interface RaktarRepository  extends JpaRepository<Raktar, String> {
+public interface RaktarRepository extends JpaRepository<Raktar, Long> {
 }
